Add openmetrics format for prometheus support
